// services/configd/tests/generated/v1/configd.ts
export interface ConsumerBinding {
  placement: PlacementBinding | undefined;
  consumerId: string;
  purpose: string;
}

export interface PlacementBinding {
  placementId: string;
  global?: GlobalPlacementScope | undefined;
  tenant?: TenantPlacementScope | undefined;
  workspace?: WorkspacePlacementScope | undefined;
  user?: UserPlacementScope | undefined;
}

export interface GlobalPlacementScope {
}

export interface TenantPlacementScope {
  tenantId: string;
}

export interface WorkspacePlacementScope {
  tenantId: string;
  workspaceId: string;
}

export interface UserPlacementScope {
  tenantId: string;
  accountPrincipalId: string;
}
// ...
export const ConsumerBinding: MessageFns<ConsumerBinding> = {
  encode(message: ConsumerBinding, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.placement !== undefined) {
      PlacementBinding.encode(message.placement, writer.uint32(10).fork()).join();
    }
    if (message.consumerId !== "") {
      writer.uint32(18).string(message.consumerId);
    }
    if (message.purpose !== "") {
      writer.uint32(26).string(message.purpose);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ConsumerBinding {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseConsumerBinding();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.placement = PlacementBinding.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.consumerId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.purpose = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ConsumerBinding {
    return {
      placement: isSet(object.placement) ? PlacementBinding.fromJSON(object.placement) : undefined,
      consumerId: isSet(object.consumerId) ? globalThis.String(object.consumerId) : "",
      purpose: isSet(object.purpose) ? globalThis.String(object.purpose) : "",
    };
  },

  toJSON(message: ConsumerBinding): unknown {
    const obj: any = {};
    if (message.placement !== undefined) {
      obj.placement = PlacementBinding.toJSON(message.placement);
    }
    if (message.consumerId !== "") {
      obj.consumerId = message.consumerId;
    }
    if (message.purpose !== "") {
      obj.purpose = message.purpose;
    }
    return obj;
  },

  create(base?: DeepPartial<ConsumerBinding>): ConsumerBinding {
    return ConsumerBinding.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ConsumerBinding>): ConsumerBinding {
    const message = createBaseConsumerBinding();
    message.placement = (object.placement !== undefined && object.placement !== null)
      ? PlacementBinding.fromPartial(object.placement)
      : undefined;
    message.consumerId = object.consumerId ?? "";
    message.purpose = object.purpose ?? "";
    return message;
  },
};

function createBasePlacementBinding(): PlacementBinding {
  return { placementId: "", global: undefined, tenant: undefined, workspace: undefined, user: undefined };
}

export const PlacementBinding: MessageFns<PlacementBinding> = {
  encode(message: PlacementBinding, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.placementId !== "") {
      writer.uint32(10).string(message.placementId);
    }
    if (message.global !== undefined) {
      GlobalPlacementScope.encode(message.global, writer.uint32(18).fork()).join();
    }
    if (message.tenant !== undefined) {
      TenantPlacementScope.encode(message.tenant, writer.uint32(26).fork()).join();
    }
    if (message.workspace !== undefined) {
      WorkspacePlacementScope.encode(message.workspace, writer.uint32(34).fork()).join();
    }
    if (message.user !== undefined) {
      UserPlacementScope.encode(message.user, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PlacementBinding {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlacementBinding();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.placementId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.global = GlobalPlacementScope.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.tenant = TenantPlacementScope.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.workspace = WorkspacePlacementScope.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.user = UserPlacementScope.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PlacementBinding {
    return {
      placementId: isSet(object.placementId) ? globalThis.String(object.placementId) : "",
      global: isSet(object.global) ? GlobalPlacementScope.fromJSON(object.global) : undefined,
      tenant: isSet(object.tenant) ? TenantPlacementScope.fromJSON(object.tenant) : undefined,
      workspace: isSet(object.workspace) ? WorkspacePlacementScope.fromJSON(object.workspace) : undefined,
      user: isSet(object.user) ? UserPlacementScope.fromJSON(object.user) : undefined,
    };
  },

  toJSON(message: PlacementBinding): unknown {
    const obj: any = {};
    if (message.placementId !== "") {
      obj.placementId = message.placementId;
    }
    if (message.global !== undefined) {
      obj.global = GlobalPlacementScope.toJSON(message.global);
    }
    if (message.tenant !== undefined) {
      obj.tenant = TenantPlacementScope.toJSON(message.tenant);
    }
    if (message.workspace !== undefined) {
      obj.workspace = WorkspacePlacementScope.toJSON(message.workspace);
    }
    if (message.user !== undefined) {
      obj.user = UserPlacementScope.toJSON(message.user);
    }
    return obj;
  },

  create(base?: DeepPartial<PlacementBinding>): PlacementBinding {
    return PlacementBinding.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PlacementBinding>): PlacementBinding {
    const message = createBasePlacementBinding();
    message.placementId = object.placementId ?? "";
    message.global = (object.global !== undefined && object.global !== null)
      ? GlobalPlacementScope.fromPartial(object.global)
      : undefined;
    message.tenant = (object.tenant !== undefined && object.tenant !== null)
      ? TenantPlacementScope.fromPartial(object.tenant)
      : undefined;
    message.workspace = (object.workspace !== undefined && object.workspace !== null)
      ? WorkspacePlacementScope.fromPartial(object.workspace)
      : undefined;
    message.user = (object.user !== undefined && object.user !== null)
      ? UserPlacementScope.fromPartial(object.user)
      : undefined;
    return message;
  },
};

function createBaseGlobalPlacementScope(): GlobalPlacementScope {
  return {};
}

export const GlobalPlacementScope: MessageFns<GlobalPlacementScope> = {
  encode(_: GlobalPlacementScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GlobalPlacementScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGlobalPlacementScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): GlobalPlacementScope {
    return {};
  },

  toJSON(_: GlobalPlacementScope): unknown {
    const obj: any = {};
    return obj;
  },

  create(base?: DeepPartial<GlobalPlacementScope>): GlobalPlacementScope {
    return GlobalPlacementScope.fromPartial(base ?? {});
  },
  fromPartial(_: DeepPartial<GlobalPlacementScope>): GlobalPlacementScope {
    const message = createBaseGlobalPlacementScope();
    return message;
  },
};

function createBaseTenantPlacementScope(): TenantPlacementScope {
  return { tenantId: "" };
}

export const TenantPlacementScope: MessageFns<TenantPlacementScope> = {
  encode(message: TenantPlacementScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tenantId !== "") {
      writer.uint32(10).string(message.tenantId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TenantPlacementScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTenantPlacementScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tenantId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TenantPlacementScope {
    return { tenantId: isSet(object.tenantId) ? globalThis.String(object.tenantId) : "" };
  },

  toJSON(message: TenantPlacementScope): unknown {
    const obj: any = {};
    if (message.tenantId !== "") {
      obj.tenantId = message.tenantId;
    }
    return obj;
  },

  create(base?: DeepPartial<TenantPlacementScope>): TenantPlacementScope {
    return TenantPlacementScope.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TenantPlacementScope>): TenantPlacementScope {
    const message = createBaseTenantPlacementScope();
    message.tenantId = object.tenantId ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}